+++ src/types/index.ts
@@ -5,6 +5,7 @@ export type OnAuthRequiredFunction = (oktaAuth: OktaAuth) => Promise<void> | voi
 export interface OktaVueOptions {
   oktaAuth: OktaAuth;
   onAuthRequired: OnAuthRequiredFunction;
+  onAuthResume: OnAuthRequiredFunction;
 }
 
 declare module '@vue/runtime-core' {